+++ src/blob/handlers/BlobHandler.ts
@@ -45,9 +45,10 @@ export default class BlobHandler extends BaseHandler implements IBlobHandler {
     metadataStore: IBlobMetadataStore,
     extentStore: IExtentStore,
     logger: ILogger,
+    loose: boolean,
     private readonly rangesManager: IPageBlobRangesManager
   ) {
-    super(metadataStore, extentStore, logger);
+    super(metadataStore, extentStore, logger, loose);
   }
 
   public setAccessControl(
@@ -886,7 +887,7 @@ export default class BlobHandler extends BaseHandler implements IBlobHandler {
       container,
       blob,
       tier,
-      undefined
+      options.leaseAccessConditions
     );
 
     const response: Models.BlobSetTierResponse = {