+++ src/blob/handlers/BlockBlobHandler.ts
@@ -358,6 +358,7 @@ export default class BlockBlobHandler extends BaseHandler
       accountName,
       containerName,
       blobName,
+      options.snapshot,
       undefined,
       options.leaseAccessConditions
     );