+++ src/blob/utils/constants.ts
@@ -1,8 +1,8 @@
 import { StoreDestinationArray } from "../../common/persistence/IExtentStore";
 import * as Models from "../generated/artifacts/models";
 
-export const VERSION = "3.9.0";
-export const BLOB_API_VERSION = "2020-02-10";
+export const VERSION = "3.11.0";
+export const BLOB_API_VERSION = "2020-06-12";
 export const DEFAULT_BLOB_SERVER_HOST_NAME = "127.0.0.1"; // Change to 0.0.0.0 when needs external access
 export const DEFAULT_LIST_BLOBS_MAX_RESULTS = 5000;
 export const DEFAULT_LIST_CONTAINERS_MAX_RESULTS = 5000;
@@ -96,6 +96,8 @@ export const DEFAULT_BLOB_PERSISTENCE_ARRAY: StoreDestinationArray = [
 ];
 
 export const ValidAPIVersions = [
+  "2020-06-12",
+  "2020-04-08",
   "2020-02-10",
   "2019-12-12",
   "2019-10-10",